refactor: Consolidate WebSocket command handling and subscription; remove unused command and event modules
