+++ src/handler/mod.rs
@@ -1,3 +1,2 @@
-pub mod command;
-pub mod event;
 pub mod message;
+pub mod types;
